third_party/blink/web_tests/fast/forms/week-multiple-fields/week-multiple-fields-blur-and-focus-events-expected.txt

Check blur and focus events for multiple fields week input UI

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

focus() and blur()
PASS testInput.focus(); state() is "blur=0 focus=1"
PASS testInput.blur(); state() is "blur=1 focus=1"
focus and Tab key to blur
PASS keyDown("	"); state() is "blur=0 focus=1"
PASS keyDown("	"); state() is "blur=0 focus=1"
PASS keyDown("	"); state() is "blur=0 focus=1"
PASS keyDown("	"); state() is "blur=1 focus=1"


PASS successfullyParsed is true

TEST COMPLETE
